+++ Spec/RealtimeClient.swift
@@ -6,7 +6,7 @@
 //  Copyright (c) 2015 Ably. All rights reserved.
 //
 
-import AblyRealtime
+import Ably
 import Quick
 import Nimble
